pyproject.toml
[project]
name = "shadowmere"
version = "1.0.0"
description = "An automatically tested list of Shadowsocks proxies."
authors = ["Your Name <you@example.com>"]
license = "GPLv3"
readme = "README.md"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0"
celery = "^5.3.6"
django-minio-storage = "^0.5.7"
django-prometheus = "^2.3.1"
django-storages = "^1.14.2"
djangorestframework = "^3.14.0"
emoji-country-flag = "^1.3.2"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
qrcode = "^7.4.2"
requests = "^2.31.0"
redis = "^5.0.1"
sentry-sdk = "^1.38.0"
django-admin-rangefilter = "^0.12.0"
django-filter = "^23.5"
django-import-export = "^3.3.3"
botocore = "^1.33.11"
django-environ = "^0.11.2"
uvicorn = "^0.24.0.post1"
gunicorn = "^21.2.0"
django-jazzmin = "^2.6.0"
huey = "^2.5.0"
humanfriendly = "^10.0"
django-ratelimit = "^4.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"
ruff-lsp = "^0.0.45"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["migrations"]
select = ["E", "F", "I", "W", "N"]
ignore = []
line-length = 120
target-version = "py312"
fix = true

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 120

[tool.ruff.lint.isort]
combine-as-imports = true
